requirements.txt
flask
python-dotenv
textblob
spacy
